Move avalanche simulator to integration tests
